test: move tests for `LengthOrPercentage`
Move tests for `LengthOrPercentage` into relevant file
