+++ src/parser/length_or_percentage.rs
@@ -39,3 +39,40 @@ impl CanStart for LengthOrPercentage {
         )
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::{length::LengthUnit, *};
+
+    fn parse_length_or_percentage(input: &str) -> Result<LengthOrPercentage, ParsingError> {
+        let mut parser = Parser::new(input.chars());
+        return parser.parse::<LengthOrPercentage>();
+    }
+
+    #[test]
+    fn nothing() {
+        assert!(parse_length_or_percentage("").is_err());
+    }
+
+    #[test]
+    fn percentage() {
+        let mut parser = Parser::new("89%".chars());
+        let result = parser.parse::<LengthOrPercentage>();
+
+        assert_eq!(result, Ok(LengthOrPercentage::Percentage(Percentage(89.0))));
+    }
+
+    #[test]
+    fn length() {
+        let mut parser = Parser::new("35px".chars());
+        let result = parser.parse::<LengthOrPercentage>();
+
+        assert_eq!(
+            result,
+            Ok(LengthOrPercentage::Length(Length::Length(
+                35.0,
+                LengthUnit::Pixels
+            )))
+        );
+    }
+}
